lift login to app component, share to children
