+++ pages/_app.tsx
@@ -6,6 +6,13 @@ import Bottom from "../components/bottom";
 import UserPrefs from "../components/userPrefs";
 import ScreenWidth from "../components/ScreenWidth";
 import { useRouter } from 'next/router'
+import { CognitoJwtVerifier } from "aws-jwt-verify";
+
+const verifier = CognitoJwtVerifier.create({
+    userPoolId: process.env.NEXT_PUBLIC_cognito_userpool_id,
+    clientId: process.env.NEXT_PUBLIC_cognito_client_id,
+    tokenUse: "access"
+})
 
 export const LoginInfo = createContext(null)
 
@@ -13,11 +20,50 @@ export function useLoginInfoContext() {
     return useContext(LoginInfo)
 }
 
+async function checkLogin(token, setLogin) {
+    if (typeof window !== 'undefined') {
+        if (token && token !== null) {
+            try {
+                await verifier.verify(token);
+                setLogin(true)
+            } catch {
+                console.log("Token not valid!");
+            }
+        }
+    }
+}
+
+async function getIdentity(ID, setId) {
+    if (typeof window !== 'undefined') {
+        if (ID && ID !== null) {
+            try {
+                const id_token = ID.split('.')[1]
+                const decode = atob(id_token)
+                const parsedUser = JSON.parse(decode)
+                setId(parsedUser.email)
+            } catch {
+                console.log("ID not valid");
+            }
+        }
+    }
+}
+
+
+
 export default function App({ Component, pageProps: { session, ...pageProps } }) {
     const router = useRouter()
 
     const [token, setToken] = useState<string | undefined>()
     const [id, setID] = useState<string | undefined>()
+    const [login, setLogin] = useState(false)
+    const [userName, setUserName] = useState<string | undefined>()
+
+    async function logout() {
+        localStorage.setItem("access_token", null);
+        localStorage.setItem("id_token", null);
+        setLogin(false)
+        setUserName('')
+    }
 
     useEffect(() => {
         if (localStorage.siteDarkMode === "true" || (!("siteDarkMode" in localStorage) && window.matchMedia("(prefers-color-scheme: dark)").matches)) {
@@ -56,21 +102,29 @@ export default function App({ Component, pageProps: { session, ...pageProps } })
             }, {})
 
             if (tokens?.access_token) {
-                console.log('setting tokens', tokens.access_token)
+
                 localStorage.setItem("access_token", tokens.access_token);
                 setToken(tokens.access_token)
                 localStorage.setItem("id_token", tokens.id_token);
                 setID(tokens.id_token)
-                router.push("/")
-            }
 
+                checkLogin(tokens.access_token, setLogin)
+                getIdentity(tokens.id_token, setUserName)
+
+                router.push("/")
+            } else {
+                checkLogin(localStorage.getItem('access_token'), setLogin)
+                getIdentity(localStorage.getItem('id_token'), setUserName)
+            }
 
         }
     }, [])
 
+    console.log(login, userName)
+
     return (
         <>
-            <LoginInfo.Provider value={{ id: id, token: token }} >
+            <LoginInfo.Provider value={{ id: id, token: token, login: login, userName: userName, logout: logout }} >
                 <UserPrefs />
                 <Topper />
                 <div className="font-body pt-7">
